# api/openapi.yaml
openapi: 3.1.0
info:
  title: KubeVirt Shepherd API
  description: |
    API for managing KubeVirt virtual machines and cloud-native governance.
    Adheres to ADR-0021: API Contract-First Development.
  version: 1.0.0
  contact:
    name: KubeVirt Shepherd Team
    url: https://github.com/kv-shepherd/shepherd
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /api/v1
    description: Production API
...
paths:
...
  /systems/{system_id}/members:
    get:
      tags: [systems]
      summary: List system members
      operationId: listSystemMembers
      parameters:
        - $ref: '#/components/parameters/SystemID'
      responses:
        '200':
          description: System member list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemMemberList'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [systems]
      summary: Add system member
      operationId: addSystemMember
      parameters:
        - $ref: '#/components/parameters/SystemID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemMemberCreateRequest'
      responses:
        '201':
          description: System member created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemMember'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /systems/{system_id}/member-candidates:
    get:
      tags: [systems, admin]
      summary: List addable system member candidates
      operationId: listSystemMemberCandidates
      parameters:
        - $ref: '#/components/parameters/SystemID'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/Search'
      responses:
        '200':
          description: Addable user candidates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /systems/{system_id}/members/{user_id}:
    patch:
      tags: [systems]
      summary: Update system member role
      operationId: updateSystemMemberRole
      parameters:
        - $ref: '#/components/parameters/SystemID'
        - $ref: '#/components/parameters/UserID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemMemberRoleUpdateRequest'
      responses:
        '200':
          description: System member updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemMember'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [systems]
      summary: Remove system member
      operationId: deleteSystemMember
      parameters:
        - $ref: '#/components/parameters/SystemID'
        - $ref: '#/components/parameters/UserID'
      responses:
        '204':
          description: System member removed
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Services ────────────────────────────────────────
...
  /admin/users:
    get:
      tags: [admin]
      summary: List users
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/Search'
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
    post:
      tags: [admin]
      summary: Create a local JWT user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Reusable Parameters (ADR-0023 Pagination) ──────
  parameters:
    Page:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Search:
      name: search
      in: query
      description: Case-insensitive search text
      schema:
        type: string
...
    SystemID:
      name: system_id
      in: path
      required: true
      schema:
        type: string
...
    UserID:
      name: user_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required: [code]
      properties:
        code:
          type: string
          description: Machine-readable error code (frontend handles i18n)
        message:
          type: string
          description: English log message (not for UI display)
        params:
          type: object
          additionalProperties: true
        field_errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required: [field, code]
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer

    # ── System ──────────────────────────────────────
...
    User:
      type: object
      required: [id, username, enabled, created_at]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        display_name:
          type: string
        enabled:
          type: boolean
        roles:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
    UserList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/Pagination'
    UserCreateRequest:
      type: object
      required: [username, password]
      properties:
        username:
          x-oapi-codegen-extra-tags:
            validate: "required"
          type: string
        password:
          x-oapi-codegen-extra-tags:
            validate: "required"
          type: string
          format: password
        email:
          type: string
        display_name:
          type: string
        enabled:
          type: boolean
        force_password_change:
          type: boolean
...
    SystemMember:
      type: object
      required: [user_id, username, role]
      properties:
        user_id:
          type: string
        username:
          type: string
        email:
          type: string
        display_name:
          type: string
        role:
          type: string
          enum: [owner, admin, member, viewer]
        created_at:
          type: string
          format: date-time
    SystemMemberList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SystemMember'
        pagination:
          $ref: '#/components/schemas/Pagination'
    SystemMemberCreateRequest:
      type: object
      required: [user_id, role]
      properties:
        user_id:
          x-oapi-codegen-extra-tags:
            validate: "required"
          type: string
        role:
          x-oapi-codegen-extra-tags:
            validate: "required,oneof=owner admin member viewer"
          type: string
          enum: [owner, admin, member, viewer]
    SystemMemberRoleUpdateRequest:
      type: object
      required: [role]
      properties:
        role:
          x-oapi-codegen-extra-tags:
            validate: "required,oneof=owner admin member viewer"
          type: string
          enum: [owner, admin, member, viewer]
...
security:
  - BearerAuth: []